renamed + button to start simulator
